requirements.txt
a2wsgi==1.10.10
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.2
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.7
chromadb==1.5.6
click==8.3.2
colorama==0.4.6
comtypes==1.4.16; platform_system == "Windows"
cryptography==46.0.7
deep-translator==1.11.4
distro==1.9.0
durationpy==0.10
ecdsa==0.19.2
fastapi==0.135.3
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
google-generativeai>=0.8.0
groq==1.1.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==3.2.0
hf-xet==1.4.3
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.27.0
huggingface_hub==1.9.1
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
jaraco.context==6.1.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langdetect==1.0.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
more-itertools==11.0.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
overrides==7.7.0
packaging==26.0
passlib==1.7.4
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyaudio==0.2.14; platform_system == "Windows"
PyAutoGUI==0.9.54
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.20.0
PyMsgBox==2.0.1
pyparsing==3.3.2
pdf2image==1.17.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.2.0
youtube-transcript-api==1.2.4
pyperclip==1.11.0
PyPika==0.51.1
pypiwin32==223; platform_system == "Windows"
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.24
pyttsx3==2.99
pytweening==1.2.0
pywin32==311; platform_system == "Windows"
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8.3
SpeechRecognition==3.16.0
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
waitress==3.0.2
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wolframalpha==5.1.3
xmltodict==1.0.4
zipp==3.23.0
